Clarify name handling a bit.
We essentially don't want to change the Galaxy API to loose the URL (https://github.com/galaxyproject/galaxy/issues/3695, https://github.com/galaxyproject/galaxy/pull/3996), we tried at some point and had to revert it. The new upload API is smart though and can be used to set a name at upload time and embed the URL we may want to have around later in the info of the dataset. I'd prefer to make this explicit in format if that is okay, so it remains a proper superset of the real API format - I've added an example of doing that here. Unfortunately for the legacy API, this will not work because we aren't exposing the name option in bioblend's library upload options.