Added flag to skip verification of SSL certificate
Update to common_parser to include an optional flag to skip verification of the SSL certificate when connecting to a Galaxy instance. Default value of 'False' when not included in a function call (SSL certificate will be verified). Including the flag in the function call sets the value to 'True' (SSL certificate will NOT be verified).